Validation fixed

+++ controllers/studentsController.js
@@ -1,22 +1,28 @@
 const mongodb = require("../data/database.js");
 const createObjectId = require("mongodb").ObjectId.createFromHexString;
 const { validationResult } = require("express-validator");
-const { validate } = require("../helpers/validation.js");
+const { blogSchema } = require("../helpers/validation.js");
+const createError = require("http-errors");
 
 const getstudents = async (req, res, next) => {
   //#swagger.tags=['Users']
   try {
-    const userId = createObjectId(req.params.id);
+    const _userId = req.params.id;
+    const regex = /^[a-zA-Z0-9]+$/;
+    if (_userId.length != 24) throw createError(400, "Invalid Length");
+    if (!regex.test(_userId)) throw createError(400, "Non-Alphanumeric ID");
+    const userId = createObjectId(_userId);
     const result = await mongodb
       .getDb()
       .db("project2")
       .collection("students")
-      .find({ _id: userId });
-    result.toArray().then((lists) => {
-      res.setHeader("Content-Type", "application/json");
-      res.status(200).json(lists[0]);
-    });
+      .findOne({ _id: userId });
+    if (!result) throw createError(404, "Nothing Found");
+    const resultJSON = JSON.parse(JSON.stringify(result));
+    res.setHeader("Content-Type", "application/json");
+    res.status(200).json(resultJSON);
   } catch (error) {
+    if (error.isJoi === true) error.status = 422;
     next(error);
   }
 };
@@ -34,6 +40,7 @@ const getAllstudents = async (req, res, next) => {
       res.status(200).json(lists);
     });
   } catch (error) {
+    if (error.isJoi === true) error.status = 422;
     next(error);
   }
 };
@@ -41,28 +48,29 @@ const getAllstudents = async (req, res, next) => {
 const createstudents = async (req, res, next) => {
   //#swagger.tags=['Users']
   try {
-    // Validate request body
-    validate("createstudents")(req, res, next);
-
     const students = {
       firstName: req.body.firstName,
       lastName: req.body.lastName,
       email: req.body.email,
-      studentsId: req.body.studentsId,
+      studentID: req.body.studentID,
       birthday: req.body.birthday,
     };
+
+    // Validate
+    const studentData = await blogSchema.validateAsync(students);
 
     const response = await mongodb
       .getDb()
       .db("project2")
       .collection("students")
-      .insertOne(students);
+      .insertOne(studentData);
     if (response.acknowledged) {
       res.status(200).json(response);
     } else {
       res.status(500).json(response.error || "There has been an error");
     }
   } catch (error) {
+    if (error.isJoi === true) error.status = 422;
     next(error);
   }
 };
@@ -70,29 +78,33 @@ const createstudents = async (req, res, next) => {
 const updatestudents = async (req, res, next) => {
   //#swagger.tags=['Users']
   try {
-    // Validate request parameters and body
-    validate("updatestudents")(req, res, next);
-
-    const userId = createObjectId(req.params.id);
+    const _userId = req.params.id;
+    const regex = /^[a-zA-Z0-9]+$/;
+    if (_userId.length != 24) throw createError(400, "Invalid Length");
+    if (!regex.test(_userId)) throw createError(400, "Non-Alphanumeric ID");
+    const userId = createObjectId(_userId);
     const students = {
       firstName: req.body.firstName,
       lastName: req.body.lastName,
       email: req.body.email,
-      studentsId: req.body.studentsId,
+      studentID: req.body.studentID,
       birthday: req.body.birthday,
     };
+    // Validate request parameters and body
+    const studentData = await blogSchema.validateAsync(students);
 
     const response = await mongodb
       .getDb()
       .db("project2")
       .collection("students")
-      .replaceOne({ _id: userId }, students);
+      .replaceOne({ _id: userId }, studentData);
     if (response.acknowledged) {
       res.status(200).json(response);
     } else {
-      res.status(500).json(response.error || "There has been an error");
+      throw createError(500, "Error Occurred during Update");
     }
   } catch (error) {
+    if (error.isJoi === true) error.status = 422;
     next(error);
   }
 };
@@ -100,18 +112,25 @@ const updatestudents = async (req, res, next) => {
 const deletestudents = async (req, res, next) => {
   //#swagger.tags=['Users']
   try {
-    const userId = createObjectId(req.params.id);
+    const _userId = req.params.id;
+    const regex = /^[a-zA-Z0-9]+$/;
+    if (_userId.length != 24) throw createError(400, "Invalid Length");
+    if (!regex.test(_userId)) throw createError(400, "Non-Alphanumeric ID");
+    const userId = createObjectId(_userId);
     const response = await mongodb
       .getDb()
       .db("project2")
       .collection("students")
       .deleteOne({ _id: userId });
-    if (response.acknowledged) {
+    if (response.deletedCount == 0) {
+      throw createError(404, "Nothing to delete");
+    } else if (response.deletedCount > 0) {
       res.status(200).json(response);
     } else {
-      res.status(500).json(response.error || "There has been an error");
+      throw createError(500, "Error Occurred during Deletion");
     }
   } catch (error) {
+    if (error.isJoi === true) error.status = 422;
     next(error);
   }
 };
